fix: revert "feat: ADDON-61556 Introduced checkbox group component (#394)"
This reverts commit 12652957396b5009d20c323fac041e64130db694.
