+++ src/main/webapp/constants/ControlTypeMap.ts
@@ -8,7 +8,6 @@ import RadioComponent from '../components/RadioComponent';
 import PlaceholderComponent from '../components/PlaceholderComponent';
 import CustomControl from '../components/CustomControl';
 import FileInputComponent from '../components/FileInputComponent';
-import CheckboxGroupsComponent from '../components/CheckboxGroupsComponent';
 
 export default {
     text: TextComponent,
@@ -20,6 +19,5 @@ export default {
     radio: RadioComponent,
     file: FileInputComponent,
     placeholder: PlaceholderComponent,
-    checkboxGroups: CheckboxGroupsComponent,
     custom: CustomControl,
 };
